#30 - Move participation related forms to separate package
This has already happenend to those forms related to activsits and referents and prepares for the new view-model-classes for the participation timeline.
