+++ src/main/java/de/naju/adebar/web/validation/persons/EditPersonForm.java
@@ -3,6 +3,7 @@ package de.naju.adebar.web.validation.persons;
 import de.naju.adebar.model.Email;
 import de.naju.adebar.model.PhoneNumber;
 import de.naju.adebar.web.validation.core.AddressForm;
+import de.naju.adebar.web.validation.persons.participant.EditParticipantForm;
 import org.hibernate.validator.constraints.NotEmpty;
 
 /**
